# medibuddy-backend/MediBuddy-Backend-main/openapi.yaml
openapi: 3.0.3
info:
  title: MediBuddy API
  description: |
    REST API for MediBuddy — a medication management, symptom logging, and
    consultation-summary PWA designed for low-resource settings.

    ## Base URL
    All endpoints are prefixed with `/api`.

    ## Authentication
    Most endpoints require a **Bearer JWT** in the `Authorization` header.
    Obtain tokens via the OTP → verify flow (see Auth endpoints).

    ## Response Envelope
    Every response follows:
    ```json
    { "success": true, "data": { ... }, "message": "optional", "serverTime": "ISO-8601" }
    ```
    Errors:
    ```json
    { "success": false, "error": "msg", "code": "CODE", "details": [], "serverTime": "ISO-8601" }
    ```

    ## Idempotency
    Mutating endpoints accept an `Idempotency-Key` header (UUID v4).
    The server caches responses for 24 h; replaying the same key returns the cached result.

    ## Rate Limits
    | Preset | Window | Max |
    |--------|--------|-----|
    | General | 15 min | 100 |
    | Auth | 15 min | 20 |
    | OTP | 15 min | 5 |
    | Sync | 1 min | 30 |
    | Write | 1 min | 60 |

    Rate headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `Retry-After`.

  version: 1.0.0
  contact:
    name: MediBuddy Team
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:5000/api
    description: Local development
  - url: https://medibuddy-backend-w9hf.onrender.com/api
    description: Production (Render)
  - url: http://localhost:4010
...
paths:
...
  /profiles/{profileId}/medications:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listMedications
      tags: [Medications]
      summary: List medications for a profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of medications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: createMedication
      tags: [Medications]
      summary: Create a medication
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedication"
      responses:
        "201":
          description: Medication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /profiles/{profileId}/medications/{medId}:
    parameters:
      - $ref: "#/components/parameters/profileId"
      - name: medId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getMedication
      tags: [Medications]
      summary: Get a single medication
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Medication object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    put:
      operationId: updateMedication
      tags: [Medications]
      summary: Update medication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedication"
      responses:
        "200":
          description: Medication updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    delete:
      operationId: deleteMedication
      tags: [Medications]
      summary: Delete medication (soft)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Medication deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ REMINDERS ═══════
...
  /profiles/{profileId}/caregivers:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listCaregivers
      tags: [Caregivers]
      summary: List caregivers for a profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of caregivers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: addCaregiver
      tags: [Caregivers]
      summary: Add a caregiver
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaregiver"
      responses:
        "201":
          description: Caregiver added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /profiles/{profileId}/caregivers/{caregiverId}:
    parameters:
      - $ref: "#/components/parameters/profileId"
      - name: caregiverId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateCaregiver
      tags: [Caregivers]
      summary: Update caregiver settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaregiver"
      responses:
        "200":
          description: Caregiver updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    delete:
      operationId: removeCaregiver
      tags: [Caregivers]
      summary: Remove caregiver (soft-delete)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Caregiver removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ APPOINTMENTS ═══════
  /appointments/{profileId}:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listAppointments
      tags: [Appointments]
      summary: List appointments for a profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: createAppointment
      tags: [Appointments]
      summary: Create an appointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointment"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /appointments/{profileId}/upcoming:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listUpcomingAppointments
      tags: [Appointments]
      summary: List upcoming appointments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Upcoming appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /appointments/{profileId}/{appointmentId}:
    parameters:
      - $ref: "#/components/parameters/profileId"
      - name: appointmentId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAppointment
      tags: [Appointments]
      summary: Get appointment details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Appointment object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    put:
      operationId: updateAppointment
      tags: [Appointments]
      summary: Update appointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointment"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    delete:
      operationId: deleteAppointment
      tags: [Appointments]
      summary: Cancel/delete appointment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Appointment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /appointments/{profileId}/{appointmentId}/status:
    parameters:
      - $ref: "#/components/parameters/profileId"
      - name: appointmentId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateAppointmentStatus
      tags: [Appointments]
      summary: Update appointment status
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [scheduled, completed, cancelled, no-show]
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ HEALTH METRICS ═══════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Obtain via `POST /auth/verify`. Short-lived access token (15 min).
        Use `POST /auth/refresh` with the refresh token to get a new pair.
  parameters:
    profileId:
      name: profileId
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId of the profile
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: UUID v4 — server caches responses for 24 h
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, serverTime]
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (varies per endpoint)
        message:
          type: string
        serverTime:
          type: string
          format: date-time
...
    CreateMedication:
      type: object
      required: [name, times]
      properties:
        name:
          type: string
          example: "Amlodipine"
        dosage:
          type: string
          example: "5mg"
        times:
          type: array
          items:
            type: string
          example: ["08:00", "20:00"]
        timezone:
          type: string
        channels:
          type: array
          items:
            type: string
            enum: [push, sms]
          default: [push]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        clientId:
          type: string
        clientTs:
          type: string
          format: date-time
    UpdateMedication:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        times:
          type: array
          items:
            type: string
        channels:
          type: array
          items:
            type: string
            enum: [push, sms]
        isActive:
          type: boolean
...
    CreateCaregiver:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          example: "Dr. Adebayo"
        phone:
          type: string
          example: "+2348098765432"
        relation:
          type: string
          example: "doctor"
        alertOnMissedDose:
          type: boolean
          default: true
        alertThresholdMinutes:
          type: integer
          default: 30
    UpdateCaregiver:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        alertOnMissedDose:
          type: boolean
        alertThresholdMinutes:
          type: integer
    CreateAppointment:
      type: object
      required: [doctor, dateTime]
      properties:
        doctor:
          type: string
          example: "Dr. Adebayo"
        dateTime:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in minutes
          default: 30
        location:
          type: string
        type:
          type: string
          enum: [in-person, virtual]
          default: "in-person"
        notes:
          type: string
...